backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==1.4.52
pymysql==1.0.3
cryptography   # ←加上这个
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
email-validator==1.3.1
python-dotenv==1.0.0
aiofiles==23.1.0